[MOV] mail: move the files related to chatter
This commit moves the files related to chatter from `@mail/core/web` to
`@mail/chatter`. It's needed as a preparation for using the backend chatter in
the portal.

Part of task #2828744
