+++ addons/mail/__manifest__.py
@@ -139,6 +139,7 @@ For more specific needs, you may also assign custom-defined actions
             'mail/static/src/core/web_portal/**/*',
             'mail/static/src/core/web/**/*',
             'mail/static/src/**/common/**/*',
+            'mail/static/src/**/web_portal/**/*',
             'mail/static/src/**/web/**/*',
             ('remove', 'mail/static/src/core/web/wysiwyg.js'),
             ('remove', 'mail/static/src/**/*.dark.scss'),
