+++ addons/mail/static/src/chatter/web/form_compiler.js
