+++ addons/mail/static/src/chatter/web/form_controller.js
