+++ addons/mail/static/src/chatter/web/form_renderer.js
@@ -1,7 +1,7 @@
 /* @odoo-module */
 
 import { AttachmentView } from "@mail/core/common/attachment_view";
-import { Chatter } from "@mail/core/web/chatter";
+import { Chatter } from "@mail/chatter/web_portal/chatter";
 
 import { onMounted, onWillUnmount, useState } from "@odoo/owl";
