+++ addons/mail/static/src/chatter/web/form_renderer.scss
