+++ addons/mail/static/src/chatter/web/mail_composer_form.js
