+++ addons/mail/static/src/chatter/web_portal/chatter.xml
