+++ addons/mail/static/src/chatter/web_portal/composer_patch.js
