+++ addons/mail/static/tests/chatter/web/attachment_box_tests.js
