+++ addons/mail/static/tests/chatter/web/follow_button_tests.js
