+++ addons/mail/static/tests/chatter/web/follower_list_menu_tests.js
