+++ addons/mail/static/tests/chatter/web/follower_subtype_tests.js
