+++ addons/mail/static/tests/chatter/web/follower_tests.js
