+++ addons/mail/static/tests/chatter/web/form_renderer_tests.js
